Working on getLastClient

+++ app/src/main/java/com/example/solcalc/model/estimate/Estimate.kt
@@ -2,7 +2,6 @@ package com.example.solcalc.model.estimate
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
-import androidx.room.Index
 import androidx.room.PrimaryKey
 import com.example.solcalc.model.client.Client
 
@@ -15,9 +14,9 @@ data class Estimate(
     @PrimaryKey(autoGenerate = true) var eid: Int,
     var clientID: Int,
     var noSP: Int,
-    var inSize: Int,
-    var storage: Int,
-    var billCut: Float,
-    var payback: Float,
-    var totalCost: Float
+    var inSize: Double,
+    var storage: Double,
+    var billCut: Double,
+    var payback: Double,
+    var totalCost: Double
 )
